move GameResult
