+++ test_game.py
@@ -1,4 +1,5 @@
 import pytest
+
 from game import Game, GameResult
 
 
@@ -6,13 +7,16 @@ from game import Game, GameResult
 def game():
     return Game()
 
+
 def assert_illegal_argument(game, guess_number):
     with pytest.raises(TypeError):
         game.guess(guess_number)
 
+
 @pytest.mark.parametrize("invalid_input", [None, "12", "1234", "12s", "121"])
 def test_exception_when_input(game, invalid_input):
     assert_illegal_argument(game, invalid_input)
+
 
 def assert_matched_number(result, solved, strikes, balls):
     assert result is not None
@@ -20,9 +24,11 @@ def assert_matched_number(result, solved, strikes, balls):
     assert result.strikes == strikes
     assert result.balls == balls
 
+
 def test_return_solved_result_if_matched_number(game):
     game.question = "123"
     assert_matched_number(game.guess("123"), solved=True, strikes=3, balls=0)
+
 
 def test_return_solved_result_if_unmatched_number(game):
     game.question = "123"
